Changed func DataStatus end move in util
