+++ app/src/main/java/ru/practicum/android/diploma/filter/presentation/view_model/DistrictVm.kt
@@ -12,7 +12,7 @@ import kotlinx.coroutines.launch
 import ru.practicum.android.diploma.filter.data.impl.AreaRepositoryImpl
 import ru.practicum.android.diploma.filter.domain.impl.AreaControllerImpl
 import ru.practicum.android.diploma.filter.domain.models.Area
-import ru.practicum.android.diploma.hhApi.DataStatus
+import ru.practicum.android.diploma.util.DataStatus
 import ru.practicum.android.diploma.filter.presentation.fragment.AREA_ID
 import ru.practicum.android.diploma.filter.presentation.fragment.AREA_NAME
 import ru.practicum.android.diploma.filter.presentation.fragment.DistrictScreenState
